refactor: Correct namespace and path for TickerModelCustomizer
Moved `TickerModelCustomizer` to the proper namespace and directory, changing it from `Configurations` to `Customizer`. Updated references accordingly.
